Add files via upload
Nebulas: Large clouds of ionized gases (Mutara, Paulson classes, etc.). They provide tactical cover (natural cloaking) but drain shields and disrupt sensors.
Pulsars: Neutron stars emitting deadly radiation. Navigating too close damages systems and crew.
Comets: Fast-moving objects with volumetric tails. They can be analyzed to collect rare gases.
Spatial Rifts: Tears in the spacetime fabric. They act as natural teleporters that project the ship to a random point in the galaxy.
Asteroid Fields: Rocky debris posing a physical risk. Collision damage increases with ship speed.
Space Mines: Hidden explosive devices placed by hostile factions. Detectable only via close-range scanning.
Derelict Ships: Hulls of destroyed vessels. Can be dismantled (dis) to recover components and resources.
Communication Buoys: Starfleet network nodes. Being nearby enhances long-range sensors (lrs), providing detailed data on adjacent quadrants.
Defense Platforms (Turrets): Heavily armed automated sentinels protecting strategic areas of interest.
Space Monsters: Includes the Crystalline Entity and the Space Amoeba, unique creatures that actively hunt vessels to feed on their energy.
Ion Storms: Meteorological phenomena moving through the galaxy, capable of blinding sensors and diverting ships' courses.

+++ include/game_state.h
@@ -45,7 +45,7 @@ typedef struct {
 
 typedef struct {
     /* Galaxy Data - Moved to TOP for reliable alignment and sync */
-    int g[11][11][11];          /* The Galaxy Cube (BPNBS Encoding) */
+    long long g[11][11][11];    /* The Galaxy Cube (BPNBS Encoding) */
     int z[11][11][11];          /* Scanned Map Cube */
 
     /* Coordinates */
